+++ custom_components/maintenance_supporter/frontend/strategy/chunks/chunk-CUHD4535.js
@@ -1,4 +1,4 @@
-import{a as r}from"./chunk-X7OW5OLQ.js";var a=r`
+import{a as r}from"./chunk-IZ327MR5.js";var a=r`
   ha-card { overflow: hidden; }
   .card-content {
     padding: 16px;